# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SEPLAG",
        "description": "API para gerenciamento de servidores públicos, incluindo servidores efetivos, temporários, lotações, endereços e fotografias.",
        "contact": {
            "name": "Matheo R Bonucia",
            "email": "matheorb@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/lotacoes": {
            "get": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Index lotacaoapi",
                "operationId": "f0b7c82f31ae38ad284e6f2d69ac38a3",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Store lotacaoapi",
                "operationId": "7ef7872fa9371403af91c19afcb11dcd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lotacoes/{lotacao}": {
            "get": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Show lotacaoapi",
                "operationId": "f1d22c9ced01548dd7a90bc27d3e5086",
                "parameters": [
                    {
                        "name": "lotacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Update lotacaoapi",
                "operationId": "48fc32840bbc0d61fc4c9a37bcadcd2c",
                "parameters": [
                    {
                        "name": "lotacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Destroy lotacaoapi",
                "operationId": "63f33cad84b97211324b5d9d8c4f8ec7",
                "parameters": [
                    {
                        "name": "lotacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pessoas/{pessoa}/lotacoes": {
            "get": {
                "tags": [
                    "Lotações"
                ],
                "summary": "LotacoesPorPessoa lotacaoapi",
                "operationId": "ab14dfd4df16d37ec6e4178118f602b3",
                "parameters": [
                    {
                        "name": "pessoa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/unidades-lista": {
            "get": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Unidades lotacaoapi",
                "operationId": "b9841a33f5ec28143e9953184c5aae79",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}